Basic Loops
In this assignment, you will implement two different programs that use loops. The first will take a String, and search through it to find all of the ‘x’,‘y’,‘z’ characters. The second will find all of the factors for a certain number - and check for a special condition, known as a “perfect” number.