The Wanderer
In this assignment, you will implement a simple game. Using Scanner to read from the keyboard and some simple StdDraw commands to draw a picture, your game will have a player which walks toward a target. When the player reaches the target, the game ends.
I will be posting a video to Panopto, which shows my solution to the program. I would encourage you to go watch that video first, and then come back and read this spec.