Ant
In this project, you will be creating a program which will simulate an ant chasing your mouse on a 2‐D grid. This is no ordinary ant chasing your mouse! Each square on the screen has a different color, representing its state. Eachtimethattheantstepsonatile, it changes the color‐but the(old)color also affect show the ant will move next. Key ideas that will be practiced in this assignment areusing2‐Darrays,
methodsandtheStdDraw.mouseX()andStdDraw.mouseY().