Voting, Part 1: Simple Objects
In the US, we use a simple system of election, known as “First Past the Post” - where a candidate can win an election with less than 50% of the vote. In some other countries, they use a more complex system, known as the “Alternative Vote.” I don’t know if it’s really any better, but it certainly is an interesting algorithm. You can learn about it here:
https://www.youtube.com/watch?v=3Y3jE3B8HsE
We’d like to model this type of election, and so we are going to read a data file which gives us a set of candidates and the preferences of several different voters. This Project will focus on reading the file, and then storing the data in a class, and then implementing code that lets the voters choose their favorite. In the next project, we’ll actually count the votes, add code to implement the Alternative Vote, and also (probably) allow for a more flexible file format.