Voting, Part 2: The Alternative Vote
In Project 6, we wrote a Voter class, which models a single voter in an election; the voter has a list of candidates that they prefer. The vote() method of the Voter class takes a list of candidates as a parameter, and we compare the candidates to the voter’s preferences; it returns the index of the candidate chosen. In this project, we’ll be adding to that code. First, we’ll implement a simple election - meaning we’ll write a method which asks all of the voters to vote, and then collects the vote totals and immediately announces a winner.
Then, we’ll implement the Alternative Vote - where, if no one candidate got more than 50% of the vote, we’ll remove one candidate from the list (the one which got the fewest votes), and run the election again, over and over. Please review the Project 6 spec to remind yourself what was required for that project. 
You can also review the video about the Alternative Vote here: https://www.youtube.com/watch?v=3Y3jE3B8HsE