LightGrid and Tile Puzzle
In this Project, you’ll be implementing two programs. Although they look very different, the first is designed to help you implement the second. The first program is named LightGrid. In this program, you will have a 4x4 array of double, which you will draw as a grid with StdDraw. The program will continually watch for the location of the mouse; where the mouse is, it will set that element of the array to 255. All of the rest of the elements of the array will exponentially decline - at a rate of 1% every 10 milliseconds. The picture below shows LightGrid as it runs. The 255 shows the current location of the mouse (sorry, my software didn’t include the mouse in the pic-
ture). The large values represent places where the mouse has been recently; smaller values represent places that have been declining for longer.