Robots, part 2
This project is an extension of Project 9. In this project, you’ll be adding features to the Robots game which make it more playable. Key amongst these new features are teleports and a level system. Along the way, we’ll be making some structural upgrades to the design of
your progam. You’ll be using a constructor, and private variables; you will also be adding the ability to support multiple board sizes.
Throughout this spec, you should assume that all of the requirements from Project 9 still apply - unless I specifically say otherwise.